+++ packages/crayons-core/src/components/datepicker/datepicker.tsx
@@ -226,7 +226,7 @@ export class Datepicker {
   /**
    * The format of time picker .
    */
-  @Prop() timeFormat: string;
+  @Prop({ mutable: true }) timeFormat: string;
   /**
    *   Triggered when the update button clicked
    */
@@ -418,7 +418,7 @@ export class Datepicker {
         this.updateValueAndEmitEvent(e);
       }
     }
-    if (e.path[0].innerText === cancelText && !this.value) {
+    if (e.composedPath()[0].innerText === cancelText && !this.value) {
       if (this.mode === 'range') {
         this.startDate = this.endDate = undefined;
       } else {
@@ -426,15 +426,15 @@ export class Datepicker {
       }
     }
 
-    if (e.path[0].innerText === cancelText) {
+    if (e.composedPath()[0].innerText === cancelText) {
       this.handlePopoverClose(e);
     }
 
     // Close datepicker only for fwClick event of Update and cancel buttons. Since this will
     // be triggered for month and year select dropdown as well the below check is added.
     if (
-      e.path[0].innerText === updateText ||
-      e.path[0].innerText === cancelText
+      e.composedPath()[0].innerText === updateText ||
+      e.composedPath()[0].innerText === cancelText
     ) {
       this.showDatePicker = false;
       this.host.shadowRoot.querySelector('fw-popover').hide();
@@ -449,7 +449,7 @@ export class Datepicker {
     e.stopImmediatePropagation();
     if (e.composedPath()[0].classList.value.includes('range-date-input')) {
       // Range input
-      const val = e.path[0].value;
+      const val = e.composedPath()[0].value;
 
       if (!val) {
         this.value = undefined;
@@ -516,7 +516,7 @@ export class Datepicker {
         this.toMonth === 0 ? this.yearCalculation(this.year, 1) : this.year;
     } else {
       // Single Date input
-      const val = e.path[0].value;
+      const val = e.composedPath()[0].value;
 
       if (!val) {
         this.value = undefined;
@@ -577,7 +577,7 @@ export class Datepicker {
    */
   @Listen('fwChange')
   handleMonthYearDropDownSelection(e) {
-    if (e.path[0].tagName !== 'FW-DATEPICKER') {
+    if (e.composedPath()[0].tagName !== 'FW-DATEPICKER') {
       e.stopImmediatePropagation();
     }