+++ packages/crayons-core/src/components/form-control/form-control.tsx
@@ -7,6 +7,11 @@ import { hasSlot } from '../../utils';
 import { TranslationController } from '../../global/Translation';
 
 const NATIVE_CONTROLS = ['input', 'select', 'textarea'];
+
+/**
+ * @parent form
+ */
+
 @Component({
   tag: 'fw-form-control',
   styleUrl: 'form-control.scss',
@@ -30,6 +35,7 @@ export class FormControl {
     | 'URL'
     | 'TEL'
     | 'TIME'
+    | 'DATE_TIME'
     | 'RELATIONSHIP' = 'TEXT';
   @Prop({ reflect: true })
   name: any;
@@ -58,6 +64,11 @@ export class FormControl {
   touched = false;
   @Prop()
   error = '';
+  /**
+   * Prop to determine whether to render the form-control or not.
+   * Default to true.
+   */
+  @Prop() shouldRender = true;
   private slotElement;
   private crayonsControlRef;
 
@@ -154,6 +165,34 @@ export class FormControl {
         }
         break;
 
+      case 'DATE_TIME':
+        {
+          const componentProps = {
+            ...this.fieldProps,
+            name: this.name,
+            placeholder: this.placeholder,
+            label: this.label,
+            required: this.required,
+            ...this.controlProps?.inputProps(
+              this.name,
+              this.type?.toLowerCase()
+            ),
+            state: (this.touched && this.error && 'error') || 'normal',
+            ['hint-text']: this.hint,
+            ['error-text']: TranslationController.t(this.error, {
+              field: this.label || this.name,
+            }),
+            showTimePicker: true,
+          };
+          cmp = (
+            <fw-datepicker
+              {...componentProps}
+              ref={(el) => (this.crayonsControlRef = el)}
+            ></fw-datepicker>
+          );
+        }
+        break;
+
       case 'CHECKBOX':
         {
           const componentProps = {
@@ -234,6 +273,8 @@ export class FormControl {
             this.type?.toLowerCase()
           );
 
+          const fieldOptions = this.fieldProps?.field_options;
+
           let componentProps = {
             ...this.fieldProps,
             name: this.name,
@@ -253,7 +294,12 @@ export class FormControl {
             ...controlProps,
             options: this.choices,
           };
-
+          // This is to handle formserv payload which might contain a field_options object, which has parameters, option_value_path and option_label_path,
+          // that denotes which property of choices object(form schema) needs to be displayed as label and which should be stored in the backend as value
+          if (fieldOptions?.option_value_path)
+            componentProps['optionValuePath'] = fieldOptions.option_value_path;
+          if (fieldOptions?.option_label_path)
+            componentProps['optionLabelPath'] = fieldOptions.option_label_path;
           cmp = (
             <fw-select
               {...componentProps}
@@ -269,6 +315,8 @@ export class FormControl {
             this.name,
             this.type?.toLowerCase()
           );
+
+          const fieldOptions = this.fieldProps?.field_options;
 
           const componentProps = {
             ...this.fieldProps,
@@ -300,6 +348,11 @@ export class FormControl {
           }
           componentProps.noDataText =
             TranslationController.t('search.startTyping');
+
+          if (fieldOptions?.option_value_path)
+            componentProps['optionValuePath'] = fieldOptions.option_value_path;
+          if (fieldOptions?.option_label_path)
+            componentProps['optionLabelPath'] = fieldOptions.option_label_path;
 
           cmp = (
             <fw-select
@@ -365,33 +418,35 @@ export class FormControl {
 
   render(): JSX.Element {
     return (
-      <div class='form-control-container'>
-        {this.renderControl()}
-        {this.hasSlot && (
-          <label
-            htmlFor={this.name}
-            class={{
-              label: true,
-              required: this.required,
-            }}
-          >
-            {this.label}
-          </label>
-        )}
-        <slot onSlotchange={() => this.handleSlotChange()}></slot>
-        {this.hasSlot && !(this.touched && this.error) && (
-          <div class='hint' id={`hint-${this.name}`}>
-            {this.hint}
-          </div>
-        )}
-        {this.hasSlot && this.touched && this.error && (
-          <div class='error' id={`error-${this.name}`}>
-            {TranslationController.t(this.error, {
-              field: this.label || this.name,
-            })}
-          </div>
-        )}
-      </div>
+      this.shouldRender && (
+        <div class='form-control-container'>
+          {this.renderControl()}
+          {this.hasSlot && (
+            <label
+              htmlFor={this.name}
+              class={{
+                label: true,
+                required: this.required,
+              }}
+            >
+              {this.label}
+            </label>
+          )}
+          <slot onSlotchange={() => this.handleSlotChange()}></slot>
+          {this.hasSlot && !(this.touched && this.error) && (
+            <div class='hint' id={`hint-${this.name}`}>
+              {this.hint}
+            </div>
+          )}
+          {this.hasSlot && this.touched && this.error && (
+            <div class='error' id={`error-${this.name}`}>
+              {TranslationController.t(this.error, {
+                field: this.label || this.name,
+              })}
+            </div>
+          )}
+        </div>
+      )
     );
   }
 }