+++ packages/crayons-core/src/components/form/form.tsx
@@ -28,6 +28,8 @@ import {
   generateDynamicValidationSchema,
   serializeForm,
   translateErrors,
+  getMappedSchema,
+  LEGO,
 } from './form-util';
 import { debounce } from '../../utils';
 
@@ -76,12 +78,38 @@ export class Form {
    */
   @Prop() formId = uuidv4();
 
+  /**
+   * Mapper Type - LEGO | FORMSERV | CUSTOM.
+   * Defaults to `LEGO`.
+   */
+  @Prop() mapperType: 'LEGO' | 'FORMSERV' | 'CUSTOM' = LEGO;
+
+  /**
+      * A custom type mapper object that maps the type of your fields in the schema to the Internal Field Types.
+      * Internal Field Types are `TEXT`, `DROPDOWN`, `EMAIL` etc.
+      * In the example below, `1` is the type of a field in your schema
+      * that needs to correspond to `TEXT` type.
+      * Please pass include the mapper for all the field types that you want to support.
+      * Example typeMapper object : {
+             'CUSTOM_TEXT': { type: 'TEXT' },
+             'SELECT': { type: 'DROPDOWN' },
+             'TEL': { type: 'PHONE_NUMBER' },
+             'CHECKBOX': { type: 'CHECKBOX' },
+             'TEXTAREA': { type: 'PARAGRAPH' },
+             'DATETIME': { type: 'DATE_TIME' },
+             'INTEGER': { type: 'NUMBER' },
+           }
+      */
+  @Prop() customTypeMapper: any = {};
+
   @State() values: FormValues = {} as any;
   @State() touched: FormTouched<FormValues> = {} as any;
   @State() errors: FormErrors<FormValues> = {} as any;
 
   @State() formValidationSchema;
   @State() formInitialValues;
+
+  @State() formSchemaState = this.formSchema;
 
   /**
    * fwFormValuesChanged - event that gets emitted when values change.
@@ -111,18 +139,15 @@ export class Form {
       this.handleInput
     );
 
-    await this.handleFormSchemaAndInitialValuesChange(
-      this.formSchema,
-      this.initialValues
-    );
+    await this.handleSchemaPropsChange();
   }
 
   @Watch('formSchema')
-  async formSchemaHandler(formSchema) {
-    await this.handleFormSchemaAndInitialValuesChange(
-      formSchema,
-      this.initialValues
-    );
+  @Watch('mapperType')
+  @Watch('customTypeMapper')
+  async schemaPropsChangeHandler() {
+    this.controls = null;
+    await this.handleSchemaPropsChange();
   }
 
   @Watch('initialValues')
@@ -138,6 +163,20 @@ export class Form {
     this.fwFormValuesChanged.emit({
       value: values,
     });
+  }
+
+  async handleSchemaPropsChange() {
+    const newSchema = getMappedSchema({
+      type: this.mapperType,
+      schema: this.formSchema,
+      customTypeMapper: this.customTypeMapper,
+    });
+
+    this.formSchemaState = newSchema;
+    await this.handleFormSchemaAndInitialValuesChange(
+      newSchema,
+      this.initialValues
+    );
   }
 
   async handleFormSchemaAndInitialValuesChange(formSchema, initialValues) {
@@ -217,7 +256,7 @@ export class Form {
 
     let serializedValues = { ...this.values };
 
-    if (this.formSchema && Object.keys(this.formSchema).length > 0) {
+    if (this.formSchemaState && Object.keys(this.formSchemaState).length > 0) {
       serializedValues = serializeForm(serializedValues, this.fields);
     }
 
@@ -357,8 +396,9 @@ export class Form {
     const error = this.errors[control.name];
     const touched = this.touched[control.name];
     control.controlProps = this.composedUtils();
-    control.error = error || '';
+    control.error = error ?? '';
     control.touched = touched || false;
+    control.shouldRender = this.shouldRenderFormControl(control);
   }
 
   private composedUtils = (): FormUtils => {
@@ -377,8 +417,8 @@ export class Form {
     const selectProps = (field: string, inputType) => ({
       value:
         inputType === 'multi_select'
-          ? this.values[field] || []
-          : this.values[field] || '',
+          ? this.values[field] ?? []
+          : this.values[field] ?? '',
     });
 
     const formProps: FormProps = {
@@ -396,6 +436,20 @@ export class Form {
     };
   };
 
+  private shouldRenderFormControl = (control) => {
+    const type = control?.type;
+    const isValidType = type !== '' && type !== null && type !== undefined;
+    const shouldRender = isValidType;
+    return shouldRender;
+  };
+
+  /**
+   * Method to set value on the form field.
+   *
+   * @param field - name of the form field
+   * @param value - value of the form field
+   * @param shouldValidate - should this form field be validated with the updated value
+   */
   @Method()
   async setFieldValue(
     field: string,
@@ -410,6 +464,11 @@ export class Form {
     }
   }
 
+  /**
+   * Method to set errors on the form fields.
+   *
+   * @param errorObj - key value pair of [fieldName]: ErrorMessage
+   */
   @Method()
   async setFieldErrors(errorObj: FormErrors<FormValues>): Promise<void> {
     Object.entries(errorObj)?.forEach(([field, value]) => {
@@ -431,21 +490,33 @@ export class Form {
   async getValues() {
     let serializedValues: FormValues = { ...this.values };
 
-    if (this.formSchema && Object.keys(this.formSchema).length > 0) {
+    if (this.formSchemaState && Object.keys(this.formSchemaState).length > 0) {
       serializedValues = serializeForm(serializedValues, this.fields);
     }
 
     return { values: this.values, serializedValues };
   }
 
+  /**
+   *
+   * @param event : An event which takes place in the DOM
+   *
+   * Method to submit the form
+   */
   @Method()
-  async doSubmit(e) {
-    return this.handleSubmit(e);
+  async doSubmit(event?): Promise<FormSubmit> {
+    return this.handleSubmit(event);
   }
 
+  /**
+   *
+   * @param event - An event which takes place in the DOM
+   *
+   * Method to reset the form
+   */
   @Method()
-  async doReset(e) {
-    this.handleReset(e);
+  async doReset(event?): Promise<void> {
+    this.handleReset(event);
   }
 
   render() {
@@ -453,23 +524,26 @@ export class Form {
 
     return (
       <form id={`form-${this.formId}`} {...utils.formProps}>
-        {this.formSchema && Object.keys(this.formSchema).length > 0 ? (
-          this.formSchema?.fields
+        {this.formSchemaState &&
+        Object.keys(this.formSchemaState).length > 0 ? (
+          this.formSchemaState?.fields
             ?.sort((a, b) => a.position - b.position)
             .map((field) => {
               return (
-                <fw-form-control
-                  key={field.name}
-                  name={field.name}
-                  type={field.type}
-                  label={field.label}
-                  required={field.required}
-                  hint={field.hint}
-                  placeholder={field.placeholder}
-                  choices={field.choices}
-                  fieldProps={field}
-                  controlProps={utils}
-                ></fw-form-control>
+                this.shouldRenderFormControl(field) && (
+                  <fw-form-control
+                    key={field.name}
+                    name={field.name}
+                    type={field.type}
+                    label={field.label}
+                    required={field.required}
+                    hint={field.hint}
+                    placeholder={field.placeholder}
+                    choices={field.choices}
+                    fieldProps={field}
+                    controlProps={utils}
+                  ></fw-form-control>
+                )
               );
             })
         ) : (