+++ packages/crayons-core/src/components/options-list/list-options.e2e.ts
@@ -1,5 +1,44 @@
 import { newE2EPage } from '@stencil/core/testing';
 describe('fw-list-options', () => {
+  const props = {
+    options: [
+      {
+        name: 'Angela Smith',
+        subText: 'angela.smith@gmail.com',
+        email: 'angela.smith@gmail.com',
+        graphicsProps: {
+          image:
+            'https://images.unsplash.com/photo-1529778873920-4da4926a72c2?ixlib=rb-1.2.1&auto=format&fit=crop&w=300&q=80',
+        },
+      },
+      {
+        name: 'Freshdesk support from India and Berlin',
+        subText: 'support.india@freshdesk.com',
+        email: 'support.india@freshdesk.com',
+        graphicsProps: {
+          image:
+            'https://images.unsplash.com/photo-1529778873920-4da4926a72c2?ixlib=rb-1.2.1&auto=format&fit=crop&w=300&q=80',
+        },
+      },
+      {
+        name: 'Angela from Freshdesk',
+        subText: 'angela@freshdesk.in',
+        email: 'angela@freshdesk.in',
+        graphicsProps: {
+          image:
+            'https://images.unsplash.com/photo-1529778873920-4da4926a72c2?ixlib=rb-1.2.1&auto=format&fit=crop&w=300&q=80',
+        },
+      },
+    ],
+    optionLabelPath: 'name',
+    optionValuePath: 'email',
+    max: 1,
+  };
+
+  const getInnerNode = (element, query) => {
+    return element.shadowRoot.querySelector(query);
+  };
+
   it('renders', async () => {
     const page = await newE2EPage();
 
@@ -69,5 +108,182 @@ describe('fw-list-options', () => {
     // @ts-ignore: Property does not exist on type Element
     expect(creatableOption.innerText).toBe('Add "text" as a recipient');
     jest.useRealTimers();
+  });
+
+  it('Renders all the select options when optionLabelPath and optionValuePath is passed', async () => {
+    const page = await newE2EPage();
+
+    await page.setContent(`<fw-list-options>
+                            </fw-list-options>`);
+    await page.$eval(
+      'fw-list-options',
+      (elm: any, { options, optionLabelPath, optionValuePath }) => {
+        elm.options = options;
+        elm.optionLabelPath = optionLabelPath;
+        elm.optionValuePath = optionValuePath;
+      },
+      props
+    );
+    await page.waitForChanges();
+    const options = await page.findAll('fw-list-options >>> fw-select-option');
+    expect(options.length).toBe(3);
+    await options.map(async (option, index) => {
+      expect(
+        option.shadowRoot.querySelector('.description-text')['innerText']
+      ).toBe(props.options[index]['name']);
+      expect(
+        option.shadowRoot.querySelector('.description-subText')['innerText']
+      ).toBe(props.options[index]['email']);
+    });
+  });
+
+  it('renders select option with the option provided with search when optionLabelPath and optionValuePath are passed as props', async () => {
+    await jest.useFakeTimers();
+    const page = await newE2EPage();
+    await page.setContent('<fw-list-options></fw-list-options>');
+    await page.$eval(
+      'fw-list-options',
+      (elm: any, { optionLabelPath, optionValuePath }) => {
+        elm.search = () => {
+          return new Promise((resolve) => {
+            return resolve([
+              {
+                name: 'Angela Smith',
+                email: 'angela.smith@gmail.com',
+              },
+            ]);
+          });
+        };
+        elm.debounceTimer = 0;
+        elm.optionLabelPath = optionLabelPath;
+        elm.optionValuePath = optionValuePath;
+        elm.filterText = 'text';
+      },
+      props
+    );
+    await page.waitForChanges();
+    await page.waitForChanges();
+    jest.runAllTimers();
+    await page.waitForChanges();
+    await page.waitForChanges();
+    const options = await page.findAll('fw-list-options >>> fw-select-option');
+    expect(options.length).toBe(1);
+    const selectOption = await options[0].shadowRoot.querySelector(
+      '.select-option'
+    );
+    expect(selectOption['innerText']).toBe('Angela Smith');
+    jest.useRealTimers();
+  });
+
+  it('emits fwChange with selected option when optionLabelPath and optionValuePath is passed', async () => {
+    const page = await newE2EPage();
+
+    await page.setContent(`<fw-list-options>
+                            </fw-list-options>`);
+    const fwChange = await page.spyOnEvent('fwChange');
+    const element = await page.find('fw-list-options');
+    await page.$eval(
+      'fw-list-options',
+      (elm: any, { options, optionLabelPath, optionValuePath }) => {
+        elm.options = options;
+        elm.optionLabelPath = optionLabelPath;
+        elm.optionValuePath = optionValuePath;
+      },
+      props
+    );
+    await page.waitForChanges();
+    const options = await page.findAll('fw-list-options >>> fw-select-option');
+    expect(options.length).toBe(3);
+    options[1].click();
+    await page.waitForChanges();
+    expect(fwChange).toHaveReceivedEventDetail({
+      value: props.options[1]['email'],
+      meta: {
+        selectedOptions: [
+          {
+            allowDeselect: true,
+            checkbox: false,
+            disabled: false,
+            hideTick: false,
+            selected: true,
+            variant: 'standard',
+            ...props.options[1],
+          },
+        ],
+      },
+    });
+  });
+
+  it('disable the non selected options when number of selected options exceeds max value when selected options are not passed', async () => {
+    const page = await newE2EPage();
+
+    await page.setContent(`<fw-list-options multiple>
+                            </fw-list-options>`);
+    await page.$eval(
+      'fw-list-options',
+      (elm: any, { options, optionLabelPath, optionValuePath, max }) => {
+        elm.options = options;
+        elm.optionLabelPath = optionLabelPath;
+        elm.optionValuePath = optionValuePath;
+        elm.max = max;
+      },
+      props
+    );
+    await page.waitForChanges();
+    const options = await page.findAll('fw-list-options >>> fw-select-option');
+    expect(options.length).toBe(3);
+    options[0].click();
+    await page.waitForChanges();
+    expect(getInnerNode(options[0], '.select-option')).toHaveClass('selected');
+    expect(getInnerNode(options[0], '.select-option')).not.toHaveClass(
+      'disabled'
+    );
+
+    expect(getInnerNode(options[1], '.select-option')).not.toHaveClass(
+      'selected'
+    );
+    expect(getInnerNode(options[1], '.select-option')).toHaveClass('disabled');
+
+    expect(getInnerNode(options[2], '.select-option')).not.toHaveClass(
+      'selected'
+    );
+    expect(getInnerNode(options[2], '.select-option')).toHaveClass('disabled');
+  });
+
+  it('disable the non selected options when number of selected options exceeds max value when selected options are passed', async () => {
+    const page = await newE2EPage();
+
+    await page.setContent(`<fw-list-options multiple>
+                            </fw-list-options>`);
+    await page.$eval(
+      'fw-list-options',
+      (elm: any, { options, optionLabelPath, optionValuePath, max }) => {
+        elm.options = options;
+        elm.optionLabelPath = optionLabelPath;
+        elm.optionValuePath = optionValuePath;
+        elm.value = [options[0].email];
+        elm.max = max;
+      },
+      props
+    );
+    await page.waitForChanges();
+    await page.waitForChanges();
+    const options = await page.findAll('fw-list-options >>> fw-select-option');
+    expect(options.length).toBe(3);
+    await page.waitForChanges();
+    expect(getInnerNode(options[0], '.select-option')).toHaveClass('selected');
+    expect(getInnerNode(options[0], '.select-option')).not.toHaveClass(
+      'disabled'
+    );
+
+    expect(getInnerNode(options[1], '.select-option')).not.toHaveClass(
+      'selected'
+    );
+    expect(getInnerNode(options[1], '.select-option')).toHaveClass('disabled');
+
+    expect(getInnerNode(options[2], '.select-option')).not.toHaveClass(
+      'selected'
+    );
+    expect(getInnerNode(options[2], '.select-option')).toHaveClass('disabled');
   });
 });