+++ packages/crayons-core/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.1.0-beta.7](https://github.com/freshworks/crayons/compare/@freshworks/crayons@4.1.0-beta.6...@freshworks/crayons@4.1.0-beta.7) (2023-01-09)
+
+### Bug Fixes
+
+- revert readonly changes ([390aea9](https://github.com/freshworks/crayons/commit/390aea98708702cb0ae267527cde7a0ebfb372d8))
+
+### Features
+
+- **form:** mapperType and customTypeMapper prop to support `LEGO`, `FORMSERV` and `CUSTOM` type ([#761](https://github.com/freshworks/crayons/issues/761)) ([fc6692d](https://github.com/freshworks/crayons/commit/fc6692d5a3e94f181518a78cb9a83e92042d6334))
+
+## [4.1.0-beta.6](https://github.com/freshworks/crayons/compare/@freshworks/crayons@4.1.0-beta.5...@freshworks/crayons@4.1.0-beta.6) (2023-01-02)
+
+### Bug Fixes
+
+- **btn-group:** bind onSlotChange method using arrow function ([22525bd](https://github.com/freshworks/crayons/commit/22525bd47607ce00d422aa70ad7698ec2bce6abf))
+- **btn-group:** fix typo ([e724378](https://github.com/freshworks/crayons/commit/e724378d02215afa901e7d261ca432f6c388ab0e))
+- **button-group:** fix button group styling collapse ([f2f4797](https://github.com/freshworks/crayons/commit/f2f47971291b7279ba2bcf3e1c881ec38bb61968))
+- fix issue with unselecting columns in data table ([3de414f](https://github.com/freshworks/crayons/commit/3de414f8257ceee8e40e196c0d0331b941a3b555))
+
+## [4.1.0-beta.5](https://github.com/freshworks/crayons/compare/@freshworks/crayons@4.1.0-beta.4...@freshworks/crayons@4.1.0-beta.5) (2022-12-18)
+
+### Bug Fixes
+
+- fix max limitation in list options component ([9450f5c](https://github.com/freshworks/crayons/commit/9450f5cc4246d0e01858142e6775f36828882e8c))
+- set min width for select input and update docs for dynamic updation of options ([1fe7c56](https://github.com/freshworks/crayons/commit/1fe7c568ea32a99c3fa741184dd0c1f87904410e))
+
+## [4.1.0-beta.4](https://github.com/freshworks/crayons/compare/@freshworks/crayons@4.1.0-beta.3...@freshworks/crayons@4.1.0-beta.4) (2022-12-12)
+
+### Bug Fixes
+
+- **datepicker:** avoid using non-standard event.path ([#729](https://github.com/freshworks/crayons/issues/729)) ([53a7611](https://github.com/freshworks/crayons/commit/53a76119e881eccb39cb8538ff52b364414ec314))
+
+### Features
+
+- add option label path and option value path for select and list options ([229500e](https://github.com/freshworks/crayons/commit/229500e9caf3dda2046af1aba4a0c3f050911fe8))
+- **form-control:** handle formServ field_options object in form control ([a19656e](https://github.com/freshworks/crayons/commit/a19656e9e850fd33e06c7285d6bf081e06360ed0))
+- **popover:** introduced new props to facilitate delay in open/close transition in popover component ([7176beb](https://github.com/freshworks/crayons/commit/7176beb6fb441ffbb953107f9da1a20a5849c7df))
+
+### Reverts
+
+- **form:** revert build failure test commit ([477a2f0](https://github.com/freshworks/crayons/commit/477a2f0cf451282093e9c576394b00b2fcbab51e))
+
+## [4.1.0-beta.3](https://github.com/freshworks/crayons/compare/@freshworks/crayons@4.1.0-beta.2...@freshworks/crayons@4.1.0-beta.3) (2022-11-19)
+
+### Features
+
+- **file-uploader:** adding reset and get files methods ([#717](https://github.com/freshworks/crayons/issues/717)) ([cad8109](https://github.com/freshworks/crayons/commit/cad810947091de51c5d5c4aaeaf1bfa8cf405407))
+
+## [4.1.0-beta.2](https://github.com/freshworks/crayons/compare/@freshworks/crayons@4.1.0-beta.1...@freshworks/crayons@4.1.0-beta.2) (2022-11-15)
+
+### Bug Fixes
+
+- **component:** existing selection Loop is not working in Vue Project ([#719](https://github.com/freshworks/crayons/issues/719)) ([136296d](https://github.com/freshworks/crayons/commit/136296de656f9e0d0ea1757697bc72fe728d429d))
+
 ## [4.1.0-beta.1](https://github.com/freshworks/crayons/compare/@freshworks/crayons@4.1.0-beta.0...@freshworks/crayons@4.1.0-beta.1) (2022-11-07)
 
 ### Bug Fixes